+++ tests/testthat/test_12_cosewic_tools.R
@@ -1,14 +1,14 @@
 test_that("prep_spatial()", {
-  expect_silent(s <- prep_spatial(bcch))
+  expect_silent(s <- prep_spatial(bcch, crs = crs_albers_canada()))
   expect_s3_class(s, "sf")
   expect_named(s, c("record_id", "geometry")) 
   expect_equal(nrow(s), nrow(bcch))
-  expect_equal(format(sf::st_crs(s)), "NAD83 / Statistics Canada Lambert") 
+  expect_equal(format(sf::st_crs(s)), "Canada Albers Equal Area Conic") 
 })
 
 test_that("prep_spatial() diff cols", {
   b <- dplyr::rename(bcch, sp = species_id, rec = record_id)
-  expect_silent(s <- prep_spatial(b, extra = "rec"))
+  expect_silent(s <- prep_spatial(b, extra = "rec", crs = 3347))
   expect_s3_class(s, "sf")
   expect_named(s, c("rec", "geometry")) 
   expect_equal(nrow(s), nrow(b))
@@ -16,63 +16,83 @@ test_that("prep_spatial() diff cols", {
 })
 
 test_that("cosewic_eoo()", {
-  df <- prep_spatial(bcch)
-  expect_silent(e <- cosewic_eoo(df, p = 0.95, spatial = FALSE))
+  # Lambert
+  df <- prep_spatial(bcch, crs = 3347)
+  expect_silent(e <- cosewic_eoo(df, p = 0.95, spatial = FALSE, clip = NULL))
   expect_s3_class(e, "data.frame")
   expect_named(e, "eoo")
   expect_equal(e[["eoo"]], units::set_units(1243.421, "km2"), tolerance = 0.001)
   
-  expect_silent(e <- cosewic_eoo(df, p = 0.95, spatial = TRUE))
+  expect_silent(e <- cosewic_eoo(df, p = 0.95, spatial = TRUE, clip = NULL))
   expect_s3_class(e, "sf")
   expect_equal(nrow(e), 1)
   expect_equal(as.character(sf::st_geometry_type(e)), "POLYGON")
   
-  expect_silent(e <- cosewic_eoo(df, p = 1, spatial = FALSE))
+  expect_silent(e <- cosewic_eoo(df, p = 1, spatial = FALSE, clip = NULL))
   expect_equal(e[["eoo"]], units::set_units(4861.251, "km2"), tolerance = 0.001)
+  
+  # Albers
+  df <- prep_spatial(bcch, crs = crs_albers_canada())
+  expect_silent(e <- cosewic_eoo(df, p = 0.95, spatial = FALSE, clip = NULL))
+  expect_s3_class(e, "data.frame")
+  expect_named(e, "eoo")
+  expect_equal(e[["eoo"]], units::set_units(1209.179, "km2"), tolerance = 0.001)
+  
+  expect_silent(e <- cosewic_eoo(df, p = 0.95, spatial = TRUE, clip = NULL))
+  expect_s3_class(e, "sf")
+  expect_equal(nrow(e), 1)
+  expect_equal(as.character(sf::st_geometry_type(e)), "POLYGON")
+  
+  expect_silent(e <- cosewic_eoo(df, p = 1, spatial = FALSE, clip = NULL))
+  expect_equal(e[["eoo"]], units::set_units(4728.589, "km2"), tolerance = 0.001)
 })
 
 test_that("cosewic_eoo() diff cols", {
+  # Lambert
   df <- dplyr::rename(bcch, sp = species_id, rec = record_id) |>
-    prep_spatial(extra = "rec")
-  expect_silent(e <- cosewic_eoo(df, p = 0.95, spatial = FALSE))
+    prep_spatial(extra = "rec", crs = 3347)
+  expect_silent(e <- cosewic_eoo(df, p = 0.95, spatial = FALSE, clip = NULL))
   expect_s3_class(e, "data.frame")
   expect_named(e, "eoo")
   expect_equal(e[["eoo"]], units::set_units(1243.421, "km2"), tolerance = 0.001)
   
-  expect_silent(e <- cosewic_eoo(df, p = 0.95, spatial = TRUE))
+  expect_silent(e <- cosewic_eoo(df, p = 0.95, spatial = TRUE, clip = NULL))
   expect_s3_class(e, "sf")
   expect_equal(nrow(e), 1)
   expect_equal(as.character(sf::st_geometry_type(e)), "POLYGON")
   
-  expect_silent(e <- cosewic_eoo(df, p = 1, spatial = FALSE))
+  expect_silent(e <- cosewic_eoo(df, p = 1, spatial = FALSE, clip = NULL))
   expect_equal(e[["eoo"]], units::set_units(4861.251, "km2"), tolerance = 0.001)
 })
 
 test_that("cosewic_eoo() no cols", {
+  # Lambert
   df <- dplyr::select(bcch, -"species_id") |>
     dplyr::mutate(record_id = dplyr::row_number()) |>
-    prep_spatial()
-  expect_silent(e <- cosewic_eoo(df, p = 0.95, spatial = FALSE))
+    prep_spatial(crs = 3347)
+  expect_silent(e <- cosewic_eoo(df, p = 0.95, spatial = FALSE, clip = NULL))
   expect_s3_class(e, "data.frame")
   expect_named(e, "eoo")
   expect_equal(e[["eoo"]], units::set_units(1243.421, "km2"), tolerance = 0.001)
   
-  expect_silent(e <- cosewic_eoo(df, p = 0.95, spatial = TRUE))
+  expect_silent(e <- cosewic_eoo(df, p = 0.95, spatial = TRUE, clip = NULL))
   expect_s3_class(e, "sf")
   expect_equal(nrow(e), 1)
   expect_equal(as.character(sf::st_geometry_type(e)), "POLYGON")
   
-  expect_silent(e <- cosewic_eoo(df, p = 1, spatial = FALSE))
+  expect_silent(e <- cosewic_eoo(df, p = 1, spatial = FALSE, clip = NULL))
   expect_equal(e[["eoo"]], units::set_units(4861.251, "km2"), tolerance = 0.001)
 })
 
 
 test_that("cosewic_iao()", {
-  df <- prep_spatial(bcch)
+  
+  # Lambert
+  df <- prep_spatial(bcch, crs = 3347)
   expect_silent(a <- cosewic_iao(df, 
                                  cell_size = units::set_units(2, "km"), 
                                  record = "record_id", 
-                                 spatial = FALSE))
+                                 spatial = FALSE, crs = 3347))
   expect_s3_class(a, "data.frame")
   expect_equal(a, 
                dplyr::tibble(min_record = 1, max_record = 36, median_record = 1, 
@@ -82,21 +102,44 @@ test_that("cosewic_iao()", {
   expect_silent(a <- cosewic_iao(df, 
                                  cell_size = units::set_units(2, "km"), 
                                  record = "record_id", 
-                                 spatial = TRUE))
+                                 spatial = TRUE, crs = 3347))
   expect_s3_class(a, "sf")
   expect_equal(sum(a$n_records), nrow(bcch))
   expect_equal(nrow(a), 450)
   expect_equal(unique(as.character(sf::st_geometry_type(a))), "POLYGON")
   expect_snapshot_value(a, style = "json2")
+  
+  # Albers
+  df <- prep_spatial(bcch, crs = crs_albers_canada())
+  expect_silent(a <- cosewic_iao(df, 
+                                 cell_size = units::set_units(2, "km"), 
+                                 record = "record_id", 
+                                 spatial = FALSE, crs = crs_albers_canada()))
+  expect_s3_class(a, "data.frame")
+  expect_equal(a, 
+               dplyr::tibble(min_record = 1, max_record = 35, median_record = 1, 
+                             grid_size_km = units::set_units(2, "km"), 
+                             n_occupied = 33, iao = units::set_units(132, "km2")))
+  
+  expect_silent(a <- cosewic_iao(df, 
+                                 cell_size = units::set_units(2, "km"), 
+                                 record = "record_id", 
+                                 spatial = TRUE, crs = crs_albers_canada()))
+  expect_s3_class(a, "sf")
+  expect_equal(sum(a$n_records), nrow(bcch))
+  expect_equal(nrow(a), 475)
+  expect_equal(unique(as.character(sf::st_geometry_type(a))), "POLYGON")
+  expect_snapshot_value(a, style = "json2")
 })
 
 test_that("cosewic_iao() diff cols", {
+  # Lambert
   df <- dplyr::rename(bcch, sp = species_id, rec = record_id) |>
-    prep_spatial(extra = "rec")
+    prep_spatial(extra = "rec", crs = 3347)
   expect_silent(a <- cosewic_iao(df, 
                                  cell_size = units::set_units(2, "km"), 
                                  record = "rec",
-                                 spatial = FALSE))
+                                 spatial = FALSE, crs = 3347))
   expect_s3_class(a, "data.frame")
   expect_equal(a, 
                dplyr::tibble(min_record = 1, max_record = 36, median_record = 1, 
@@ -106,7 +149,7 @@ test_that("cosewic_iao() diff cols", {
   expect_silent(a <- cosewic_iao(df, 
                                  cell_size = units::set_units(2, "km"), 
                                  record = "rec", 
-                                 spatial = TRUE))
+                                 spatial = TRUE, crs = 3347))
   expect_s3_class(a, "sf")
   expect_equal(sum(a$n_records), nrow(bcch))
   expect_equal(nrow(a), 450)
@@ -117,11 +160,11 @@ test_that("cosewic_iao() diff cols", {
 test_that("cosewic_iao() no cols", {
   df <- dplyr::select(bcch, -"species_id") |>
     dplyr::mutate(record_id = dplyr::row_number()) |>
-    prep_spatial()
+    prep_spatial(crs = 3347)
   expect_silent(a <- cosewic_iao(df, 
                                  cell_size = units::set_units(2, "km"), 
                                  record = "record_id",
-                                 spatial = FALSE))
+                                 spatial = FALSE, crs = 3347))
   expect_s3_class(a, "data.frame")
   expect_equal(a, 
                dplyr::tibble(min_record = 1, max_record = 36, median_record = 1, 
@@ -131,7 +174,7 @@ test_that("cosewic_iao() no cols", {
   expect_silent(a <- cosewic_iao(df, 
                                  cell_size = units::set_units(2, "km"), 
                                  record = "record_id", 
-                                 spatial = TRUE))
+                                 spatial = TRUE, crs = 3347))
   expect_s3_class(a, "sf")
   expect_equal(sum(a$n_records), nrow(bcch))
   expect_equal(nrow(a), 450)
@@ -141,11 +184,13 @@ test_that("cosewic_iao() no cols", {
 
 
 test_that("cosewic_ranges()", {
-  expect_silent(r <- cosewic_ranges(bcch))
+  
+  # Lambert
+  expect_silent(r <- cosewic_ranges(bcch, crs = 3347))
   expect_type(r, "list")
   expect_named(r, c("iao", "eoo"))
   
-  expect_silent(r <- cosewic_ranges(bcch, spatial = FALSE))
+  expect_silent(r <- cosewic_ranges(bcch, spatial = FALSE, crs = 3347))
   expect_s3_class(r, "data.frame")
   expect_equal(
     r, 
@@ -165,15 +210,43 @@ test_that("cosewic_ranges()", {
   expect_message(r <- cosewic_ranges(bcch[1,], spatial = FALSE), 
                  "EOO is less than IAO")
   expect_equal(r$iao, r$eoo_p95)
-})
-
-test_that("cosewic_ranges() diff cols", {
-  b <- dplyr::rename(bcch, sp = species_id, rec = record_id)
-  expect_silent(r <- cosewic_ranges(b, record = "rec", species = "sp"))
+  
+  # Albers
+  expect_silent(r <- cosewic_ranges(bcch, crs = crs_albers_canada()))
   expect_type(r, "list")
   expect_named(r, c("iao", "eoo"))
   
-  expect_silent(r <- cosewic_ranges(b, record = "rec", species = "sp", spatial = FALSE))
+  expect_silent(r <- cosewic_ranges(bcch, spatial = FALSE, crs = crs_albers_canada()))
+  expect_s3_class(r, "data.frame")
+  expect_equal(
+    r, 
+    dplyr::tibble(species_id = 14280L,
+                  n_records_total = nrow(bcch),
+                  min_record = 1, max_record = 35, median_record = 1, 
+                  grid_size_km = units::set_units(2, "km"), 
+                  n_occupied = 33, iao = units::set_units(132, "km2"),
+                  eoo_p95 = units::set_units(1209.179, "km2")), 
+    tolerance = 0.001)
+  
+  expect_error(cosewic_ranges(dplyr::select(bcch, -"latitude")),
+               "`coord_lat` and `coord_lon` must be columns in `df_db`")
+  expect_error(cosewic_ranges(dplyr::mutate(bcch, latitude = collection)),
+               "`coord_lat` and `coord_lon` must be numeric")
+  
+  expect_message(r <- cosewic_ranges(bcch[1,], spatial = FALSE), 
+                 "EOO is less than IAO")
+  expect_equal(r$iao, r$eoo_p95)
+})
+
+test_that("cosewic_ranges() diff cols", {
+  # Lambert
+  b <- dplyr::rename(bcch, sp = species_id, rec = record_id)
+  expect_silent(r <- cosewic_ranges(b, record = "rec", species = "sp", crs = 3347))
+  expect_type(r, "list")
+  expect_named(r, c("iao", "eoo"))
+  
+  expect_silent(r <- cosewic_ranges(b, record = "rec", species = "sp", 
+                                    spatial = FALSE, crs = 3347))
   expect_s3_class(r, "data.frame")
   expect_equal(
     r, 
@@ -198,13 +271,13 @@ test_that("cosewic_ranges() diff cols", {
 test_that("cosewic_ranges() no cols", {
   b <- dplyr::select(bcch, -"species_id", -"record_id")
   
-  expect_warning(r <- cosewic_ranges(b), "Column \"species_id\"") |>
+  expect_warning(r <- cosewic_ranges(b, crs = 3347), "Column \"species_id\"") |>
     expect_warning("Column \"record_id\"")
-  expect_silent(r <- cosewic_ranges(b, record = NULL, species = NULL))
+  expect_silent(r <- cosewic_ranges(b, record = NULL, species = NULL, crs = 3347))
   expect_type(r, "list")
   expect_named(r, c("iao", "eoo"))
   
-  expect_silent(r <- cosewic_ranges(b, record = NULL, species = NULL, spatial = FALSE))
+  expect_silent(r <- cosewic_ranges(b, record = NULL, species = NULL, spatial = FALSE, crs = 3347))
   expect_s3_class(r, "data.frame")
   expect_equal(
     r, 
@@ -228,9 +301,9 @@ test_that("cosewic_ranges() no cols", {
 test_that("cosewic_ranges() filter_unique", {
   
   # 95% EOO
-  expect_silent(r1 <- cosewic_ranges(bcch, spatial = FALSE))
+  expect_silent(r1 <- cosewic_ranges(bcch, spatial = FALSE, crs = 3347))
   expect_warning(
-    r2 <- cosewic_ranges(bcch, spatial = FALSE, filter_unique = TRUE),
+    r2 <- cosewic_ranges(bcch, spatial = FALSE, filter_unique = TRUE, crs = 3347),
                  "This may bias non-100% EOO calculations")
   
   expect_gt(r1$max_record, r2$max_record)
@@ -238,13 +311,15 @@ test_that("cosewic_ranges() filter_unique", {
                r2[,c("grid_size_km", "n_occupied", "iao")])
   expect_false(r1$eoo_p95 == r2$eoo_p95)
   
-  expect_silent(r1 <- cosewic_ranges(bcch, spatial = FALSE))
-  expect_warning(r2 <- cosewic_ranges(bcch, filter_unique = TRUE, spatial = FALSE),
+  expect_silent(r1 <- cosewic_ranges(bcch, spatial = FALSE, crs = 3347))
+  expect_warning(r2 <- cosewic_ranges(bcch, filter_unique = TRUE, 
+                                      spatial = FALSE, crs = 3347),
                  "This may bias non-100% EOO calculations")
   
   # Full EOO
-  expect_silent(r1 <- cosewic_ranges(bcch, eoo_p = 1, spatial = FALSE))
-  expect_warning(r2 <- cosewic_ranges(bcch, filter_unique = TRUE, eoo_p = 1, spatial = FALSE),
+  expect_silent(r1 <- cosewic_ranges(bcch, eoo_p = 1, spatial = FALSE, crs = 3347))
+  expect_warning(r2 <- cosewic_ranges(bcch, filter_unique = TRUE, eoo_p = 1, 
+                                      spatial = FALSE, crs = 3347),
                  "Filtering")
   expect_gt(r1$max_record, r2$max_record)
   expect_equal(r1[,c("grid_size_km", "n_occupied", "iao")],
@@ -254,21 +329,32 @@ test_that("cosewic_ranges() filter_unique", {
 })
 
 
+test_that("cosewic_ranges() eoo clip", {
+  ON <- rnaturalearth::ne_states("Canada") |> 
+    dplyr::filter(postal == "ON")
+  mult <- rbind(bcch, hofi)
+  
+  expect_silent(r0 <- cosewic_ranges(mult, crs = 3347))
+  expect_silent(r1 <- cosewic_ranges(mult, eoo_clip = ON, crs = 3347))
+  expect_true(all(r0$eoo$eoo_p95 > r1$eoo$eoo_p95))
+})
+
+
 test_that("cosewic_plot()", {
   
-  expect_silent(r1 <- cosewic_ranges(bcch))
+  expect_silent(r1 <- cosewic_ranges(bcch, crs = 3347))
   expect_silent(g1 <- cosewic_plot(r1))
   expect_s3_class(g1, "ggplot")
   
-  expect_silent(g2 <- cosewic_plot(r1, grid = grid_canada()))  
+  expect_silent(g2 <- cosewic_plot(r1, grid = grid_canada(crs = 3347)))  
   expect_silent(g3 <- cosewic_plot(r1, points = bcch))
-  expect_silent(g4 <- cosewic_plot(r1, grid = grid_canada(), 
+  expect_silent(g4 <- cosewic_plot(r1, grid = grid_canada(crs = 3347), 
                                    map = map_canada()))
-  expect_silent(g5 <- cosewic_plot(r1, grid = grid_canada(), 
+  expect_silent(g5 <- cosewic_plot(r1, grid = grid_canada(crs = 3347), 
                                    map = map_canada(), 
                                    title = "Black-capped Chickadees"))  
   
-  expect_silent(r2 <- cosewic_ranges(rbind(bcch, hofi)))
+  expect_silent(r2 <- cosewic_ranges(rbind(bcch, hofi), crs = 3347))
   expect_silent(g6 <- cosewic_plot(r2))
   expect_false(inherits(g6, "ggplot"))
   expect_length(g6, 2)
@@ -285,23 +371,24 @@ test_that("cosewic_plot()", {
 test_that("cosewic_plot() no cols", {
   b <- dplyr::select(bcch, -"species_id", -"record_id")
   
-  expect_silent(r1 <- cosewic_ranges(b, species = NULL, record = NULL))
+  expect_silent(r1 <- cosewic_ranges(b, species = NULL, record = NULL, crs = 3347))
   expect_warning(g0 <- cosewic_plot(r1), "Column \"species_id\" not found")
   expect_silent(g1 <- cosewic_plot(r1, species = NULL))
   expect_s3_class(g1, "ggplot")
   expect_equal(g0, g1)
   
   
-  expect_silent(g2 <- cosewic_plot(r1, grid = grid_canada(), species = NULL))  
+  expect_silent(g2 <- cosewic_plot(r1, grid = grid_canada(crs = 3347), species = NULL))  
   expect_silent(g3 <- cosewic_plot(r1, points = bcch, species = NULL))
-  expect_silent(g4 <- cosewic_plot(r1, grid = grid_canada(), 
+  expect_silent(g4 <- cosewic_plot(r1, grid = grid_canada(crs = 3347), 
                                    map = map_canada(), species = NULL))
-  expect_silent(g5 <- cosewic_plot(r1, grid = grid_canada(), 
+  expect_silent(g5 <- cosewic_plot(r1, grid = grid_canada(crs = 3347), 
                                    map = map_canada(), species = NULL,
                                    title = "Black-capped Chickadees"))  
   
   # Multiple species as one
-  expect_silent(r2 <- cosewic_ranges(rbind(bcch, hofi), species = NULL, record = NULL))
+  expect_silent(r2 <- cosewic_ranges(rbind(bcch, hofi), 
+                                     species = NULL, record = NULL, crs = 3347))
   expect_silent(g6 <- cosewic_plot(r2, species = NULL))
   expect_s3_class(g6, "ggplot")