+++ NEWS.md
@@ -1,3 +1,12 @@
+# naturecounts dev
+* Add option to clip EOO in `cosewic_ranges()` to a particular shapefile before calculating area
+* Add option to scale records in `cosewic_plot()`
+* Change API to naturecounts.ca
+* Add option to change the CRS for all COSEWIC-related functions.
+* Update default CRS for `cosweic_ranges()` to an equal area projection 
+  (add `crs_albers_canada()` to this effect)
+* Fix bug with request ids where tried to download unapproved collections.
+
 # naturecounts 0.4.1
 * Fix use of species_id and record_id in `cosewic_ranges()`.
   * Truly allow different columns